pyproject.toml
[project]
name = "pillowcases"
version = "0.1.0"
description = """\
    Library that makes it possible to work in a concise, algebraic \
    way with Python Imaging Library image objects.\
    """
license = {text = "MIT"}
authors = [
    {name = "Andrei Lapets"},
    {email = "a@lapets.io"}
]
readme = "README.rst"
requires-python = ">=3.7"
dependencies = [
    "pillow~=9.2"
]

[project.urls]
Repository = "https://github.com/lapets/pillowcases"

[project.optional-dependencies]
test = [
    "pytest~=7.0",
    "pytest-cov~=3.0"
]
publish = [
    "build~=0.8",
    "twine~=4.0"
]

[build-system]
requires = [
    "setuptools~=62.0"
]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
addopts = "--doctest-modules --ignore=docs --cov=pillowcases --cov-report term-missing"
